More 64 bit stuff
Use 64 bit sizes/counters where needed. Use rocksDB's random64.
Note that this requires compiling everything with g++ -std=c++11.

+++ doc/bench/db_bench_sqlite4.cc
@@ -169,8 +169,8 @@ class Benchmark {
  private:
   sqlite4* db_;
   int db_num_;
-  int num_;
-  int reads_;
+  int64_t num_;
+  int64_t reads_;
   double start_;
   double last_op_finish_;
   int64_t bytes_;
